maint(server): test configuration, default returns and default functions
There now exists default test configuration options that will be pulled
from the .env file for running in test mode. all response methods now
perform a return on responses made to the requesting client. And static
classes have been removed from the migration from java.
