+++ .env.example
@@ -1,9 +1,12 @@
-NODE_ENV=development
+#################################
+# Master Configuration Options  #
+#################################
 
 COOKIE_DOMAIN=
 API_KEY=
 APP_PORT=8080
 
+# Database Connection
 DB_HOST=127.0.0.1
 DB_PORT=5432
 DB_NAME=devwars
@@ -27,3 +30,13 @@ AWS_ENDPOINT_URL=
 AWS_ACCESS_KEY=
 AWS_SECRET_KEY=
 AWS_BUCKET_NAME=
+
+#################################
+# Testing Configuration Options #
+#################################
+
+TEST_DB_HOST=127.0.0.1
+TEST_DB_PORT=5432
+TEST_DB_NAME=devwars_test
+TEST_DB_USER=root
+TEST_DB_PASS=root
