+++ app/controllers/user/UserGameStats.controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
 import { getCustomRepository } from 'typeorm';
-import UserGameStats from '../../models/UserGameStats';
+
 import UserRepository from '../../repository/User.repository';
 
 export async function forUser(request: Request, response: Response) {
@@ -9,6 +9,5 @@ export async function forUser(request: Request, response: Response) {
     if (!user) return response.sendStatus(404);
 
     const stats = await userRepository.findGameStatsByUser(user);
-
-    response.json(stats);
+    return response.json(stats);
 }
