+++ app/routes/Activity.routes.ts
@@ -1,8 +1,8 @@
 import * as express from 'express';
 
+import { mustBeAuthenticated } from '../middleware/Auth.middleware';
 import * as ActivityController from '../controllers/user/Activity.controller';
 import { asyncErrorHandler } from './handlers';
-import { mustBeAuthenticated } from '../middlewares/Auth.middleware';
 
 export const ActivityRoute: express.Router = express
     .Router()
