+++ app/routes/GameSchedule.routes.ts
@@ -1,11 +1,12 @@
 import * as express from 'express';
-import * as GameScheduleController from '../controllers/game/GameSchedule.controller';
 
-import { mustBeRole, mustBeAuthenticated } from '../middlewares/Auth.middleware';
-import { UserRole } from '../models/User';
+import * as GameScheduleController from '../controllers/game/gameSchedule.controller';
+import { mustBeRole, mustBeAuthenticated } from '../middleware/Auth.middleware';
 import { asyncErrorHandler } from './handlers';
+import { UserRole } from '../models/User';
 
 const GameScheduleRoute: express.Router = express.Router();
+
 GameScheduleRoute.get('/', asyncErrorHandler(GameScheduleController.all));
 
 GameScheduleRoute.post(
