+++ app/routes/Health.route.ts
@@ -1,5 +1,10 @@
 import * as express from 'express';
-import { HealthController } from '../controllers/Health.controller';
+
+import * as HealthController from '../controllers/Health.controller';
 import { asyncErrorHandler } from './handlers';
 
-export const HealthRoute: express.Router = express.Router().get('/health', asyncErrorHandler(HealthController.index));
+const HealthRoute = express.Router();
+
+HealthRoute.get('/health', asyncErrorHandler(HealthController.index));
+
+export { HealthRoute };
