+++ app/routes/Temp.routes.ts
@@ -1,7 +1,7 @@
 import * as express from 'express';
 // TEMP: To support cookie authentication for Old Editor
-import { AuthController } from '../controllers/auth/Auth.controller';
-import { mustBeAuthenticated } from '../middlewares/Auth.middleware';
+import * as AuthController from '../controllers/authentication/authentication.controller';
+import { mustBeAuthenticated } from '../middleware/Auth.middleware';
 import { asyncErrorHandler } from './handlers';
 
 export const TempRoute: express.Router = express
