+++ config/index.ts
@@ -1,19 +1,21 @@
 import * as dotenv from 'dotenv';
 import * as AWS from 'aws-sdk';
+import { isNil } from 'lodash';
 
 dotenv.config();
 
 const DIALECT: any = process.env.DIALECT || 'postgres';
+const environment = process.env.NODE_ENV;
 
-const LOCAL_CONFIGURATION = {
-    HOST: process.env.DB_HOST,
-    PORT: process.env.DB_PORT,
-    NAME: process.env.DB_NAME,
-    USER: process.env.DB_USER,
-    PASS: process.env.DB_PASS,
+const TEST_CONFIGURATION = {
+    HOST: process.env.TEST_DB_HOST,
+    PORT: process.env.TEST_DB_PORT,
+    NAME: process.env.TEST_DB_NAME,
+    USER: process.env.TEST_DB_USER,
+    PASS: process.env.TEST_DB_PASS,
 };
 
-const PRODUCTION_CONFIGURATION = {
+const MASTER_CONFIGURATION = {
     HOST: process.env.DB_HOST,
     PORT: process.env.DB_PORT,
     NAME: process.env.DB_NAME,
@@ -22,7 +24,7 @@ const PRODUCTION_CONFIGURATION = {
 };
 
 const config = {
-    DATABASE: process.env.NODE_ENV === 'PRODUCTION' ? PRODUCTION_CONFIGURATION : LOCAL_CONFIGURATION,
+    DATABASE: !isNil(environment) && environment === 'test' ? TEST_CONFIGURATION : MASTER_CONFIGURATION,
     PORT_APP: Number(process.env.APP_PORT),
 };
